9031 - changed click fn in the DrilldownSidebar.jsx; moved isMobile detection up to StatusOfFunds.jsx and passed down needed props to VisualizationSection.jsx; removed the second back button at < 768
