+++ src/js/components/agency/statusOfFunds/StatusOfFunds.jsx
@@ -7,6 +7,8 @@ import React, { useCallback, useState, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { useSelector, useDispatch } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { tabletScreen } from 'dataMapping/shared/mobileBreakpoints';
+import { throttle } from "lodash";
 
 import { FlexGridRow, FlexGridCol, Pagination, LoadingMessage } from 'data-transparency-ui';
 import {
@@ -48,6 +50,22 @@ const StatusOfFunds = ({ fy }) => {
     const [results, setResults] = useState([]);
     const { overview, selectedSubcomponent } = useSelector((state) => state.agency);
     const [toggle, setOnToggle] = useState(false);
+    const [windowWidth, setWindowWidth] = useState(0);
+    const [isMobile, setIsMobile] = useState(window.innerWidth < tabletScreen);
+    const [viewType, setViewType] = useState(isMobile ? 'table' : 'chart');
+
+    useEffect(() => {
+        const handleResize = throttle(() => {
+            const newWidth = window.innerWidth;
+            if (windowWidth !== newWidth) {
+                setWindowWidth(newWidth);
+                setIsMobile(newWidth < tabletScreen);
+                setViewType(isMobile ? 'table' : viewType);
+            }
+        }, 50);
+        window.addEventListener('resize', handleResize);
+        return () => window.removeEventListener('resize', handleResize);
+    }, [isMobile, viewType, windowWidth]);
 
     useEffect(() => {
         if (request.current) {
@@ -178,13 +196,14 @@ const StatusOfFunds = ({ fy }) => {
     }, [fy, overview.toptierCode]);
 
     const onClick = (selectedLevel, data) => {
-    // reset to page 1 on drilldown
+        // reset to page 1 on drilldown
         setResetPageChange(true);
         const subcomponentTotalData = Object.create(BaseStatusOfFundsLevel);
         subcomponentTotalData.populate(data);
         dispatch(setSelectedSubcomponent(subcomponentTotalData));
         setSubcomponent(subcomponentTotalData);
     };
+
     const goBack = () => {
         if (overview.toptierCode) {
             setLevel(0);
@@ -216,13 +235,13 @@ const StatusOfFunds = ({ fy }) => {
                     <DrilldownSidebar
                         toggle={toggle}
                         level={level}
-                        setLevel={onClick}
+                        goBack={goBack}
                         agencyName={overview.name}
                         fy={fy}
                         selectedSubcomponent={selectedSubcomponent} />
                 </FlexGridCol>
                 <FlexGridCol className="status-of-funds__visualization" desktop={9}>
-                    {level === 1 ?
+                    {level === 1 && !isMobile ?
                         <button title="Go up a level" className="drilldown-back-button" onClick={goBack}>
                             <FontAwesomeIcon icon="arrow-left" />
                             &nbsp;&nbsp;Back
@@ -243,7 +262,10 @@ const StatusOfFunds = ({ fy }) => {
                             agencyId={overview.toptierCode}
                             agencyName={overview.name}
                             fy={fy}
-                            results={results} />
+                            results={results}
+                            isMobile={isMobile}
+                            viewType={viewType}
+                            setViewType={setViewType} />
                         :
                         <LoadingMessage /> }
                     <Pagination
